Rename docker modules to prevent existing module naming duplicate issue during pylint
